Add a simple manifest to deploy drone to cloud foundry
I don't think it's going to be possible to run the agents on CF, because
they need access to the docker socket to schedule containers, and we
can't give them that.

It is possible to run the server though, which is a start. The problem
with this is that I think the agents communicate with the server on port
9000 (not 443), which is a problem because it's only possible to expose
one port through the various load balancers and proxies that sit in
front of our apps. :thinking_face: